requirements.txt
asciinema==2.2.0
asgiref==3.5.0
Django==4.0.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
netcat==0.0.1a0
pytz==2023.3
sqlparse==0.4.2
tzdata==2022.1